Fix obs importer test
